query for unused after changes
The unused list was queried before potentially adding/removing packages,
which usually meant that you always pruned packages that were made
unnecessary *last* time. We should just query this when we need it.

While testing, I realized that `query` here would use my `pacman`
function, which is just the lovely reality of dynamic scoping. This also
failed because `dry` was undefined, since I had not added/removed
packages and was not prompted. I changed this function name to not
conflict with the command.